requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

